Frame

+++ js/actions.js
@@ -2,12 +2,12 @@
 deleteObject = document.getElementById("deleteObject");
 
 deleteObject.onclick = function () {
-    if (currentObject != null) {
+    if (currentObject != null && !isSomePointSelected) {
         deleteFunc();
     }
 }
 document.addEventListener('keydown', function (event) {
-    if (event.code == 'Delete' && currentObject != null) {
+    if (event.code == 'Delete' && currentObject != null && !isSomePointSelected) {
         deleteFunc();
     }
 });
@@ -176,7 +176,7 @@ save.onclick = function () {
         return;
     }
     if (currentObject != null) {
-        currentObject.removeFrameAndPoints();
+        currentObject.hideFrameAndPoints();
         currentObject = null;
     }
     let svgData = draw_panel.innerHTML.toString();
@@ -212,7 +212,7 @@ savePng.onclick = function () {
         return;
     }
     if (currentObject != null) {
-        currentObject.removeFrameAndPoints();
+        currentObject.hideFrameAndPoints();
         currentObject = null;
     }
     let svgData = draw_panel.innerHTML.toString();
@@ -270,8 +270,8 @@ frontObject = document.getElementById("frontObject");
 frontObject.onclick = function () {
     if (currentObject != null) {
         svgPanel.append(currentObject.svgElement);
-        for (let i = 0; i < currentObject.frame.length; i++) {
-            svgPanel.append(currentObject.frame[i].svgElement);
+        for (let i = 0; i < currentObject.frameArray.length; i++) {
+            svgPanel.append(currentObject.frameArray[i].svgElement);
         }
         for (let i = 0; i < currentObject.pointsArray.length; i++) {
             svgPanel.append(currentObject.pointsArray[i].circle);
